Good thinking Jaehwan
