+++ packages/client/hmi-client/src/components/workflow/ops/simulate-ensemble-ciemss/simulate-ensemble-util.ts
@@ -1,4 +1,3 @@
-import { getSimulation } from '@/services/models/simulation-service';
 import { EnsembleModelConfigs } from '@/types/Types';
 import { SimulateEnsembleMappingRow, SimulateEnsembleWeights } from './simulate-ensemble-ciemss-operation';
 
@@ -30,27 +29,4 @@ export function formatSimulateModelConfigurations(
 	});
 
 	return [...Object.values(ensembleModelConfigMap)];
-}
-
-// The result files from ensemble simulate have column headers such as model_#/column_name
-// This can be used to trace what model configuration is what model_# in a given run.
-export async function getResultModelConfigMap(runId: string) {
-	interface EnsembleModelConfigsSnakeCase {
-		id: string;
-		solution_mappings: { [index: string]: string };
-		weight: number;
-	}
-	const resultMap: { [key: string]: string } = {};
-
-	// Get Simulation Execution Run:
-	const simulationRun = await getSimulation(runId);
-	if (!simulationRun) {
-		console.error(`Could not find simulation ${runId}`);
-		return null;
-	}
-	const modelConfigs: EnsembleModelConfigsSnakeCase[] = simulationRun.executionPayload.model_configs;
-	for (let i = 0; i < modelConfigs.length; i++) {
-		resultMap[`model_${i}`] = modelConfigs[i].id;
-	}
-	return resultMap;
 }
